Fix for #14.
The vesselports table used a MUL index, therefore if row A exists, adding row A again would cause two copies of row A to be in the table.  As the database probes, more and more duplicates are created.  This is fixed by changing the MUL index for the port table to be a Primary index using all 3 columns.